requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
agent-framework>=1.2,<2
python-dotenv>=1,<2
